// src/model/ui/routes/InternalPortalRoutesEnum.ts
export enum InternalPortalRoutesEnum {
    SEMINARS = "/seminars",
    CALENDAR_MANAGEMENT = "/calendar-management",
    FORMS_BUILDER = "/exam-builder",
    SURVEY_BUILDER = "/survey-builder",
    REPORTS = "/reports",
    TRAINING_CAMP = "/training-camp",
    JOURNAL_OF_CANDIDATES = "/the-journal-of-initial-candidates",
    EXAMS = "/exams",
    ATTENDEE_MENTORS = "/attendee_mentors",
    COST_SHEET = "/cost_sheet",
}